+++ core-ng/src/main/java/core/framework/impl/validate/Validator.java
@@ -1,11 +1,13 @@
 package core.framework.impl.validate;
 
 import core.framework.impl.log.filter.JSONLogParam;
-import core.framework.json.JSON;
+import core.framework.internal.json.JSONMapper;
 import core.framework.util.Strings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.lang.reflect.Field;
 import java.util.Map;
 import java.util.function.Function;
@@ -24,6 +26,15 @@ public final class Validator {
         this.validator = builder.build().orElse(null);
     }
 
+    // serialize to json without any extra logic
+    private static JSONLogParam param(Object instance) {
+        try {
+            return new JSONLogParam(Strings.bytes(JSONMapper.OBJECT_MAPPER.writeValueAsString(instance)), UTF_8);
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
     public void validate(Object bean, boolean partial) {
         if (bean == null) {
             throw new ValidationException(Map.of("bean", "bean must not be null"));
@@ -33,9 +44,9 @@ public final class Validator {
             var errors = new ValidationErrors();
             validator.validate(bean, errors, partial);
             if (errors.hasError()) {
-                // all bean can be validated can be converted to JSON, only log on failure path, not slow down happy path (by toJSON)
-                // use debug level to not interfere error_code in action log, and it's logging whole bean json, not as proper error message
-                LOGGER.debug("validate, beanClass={}, bean={}, partial={}", bean.getClass().getCanonicalName(), new JSONLogParam(Strings.bytes(JSON.toJSON(bean)), UTF_8), partial);
+                // all validatable beans can be converted to JSON, only log content on failure path, not to slow down happy path
+                // use debug level not to interfere error_code in action log
+                LOGGER.debug("validate, beanClass={}, bean={}, partial={}", bean.getClass().getCanonicalName(), param(bean), partial);
                 throw new ValidationException(errors.errors);
             }
         }